pyproject.toml
[tool.poetry]
name = "xebias-airflow-workspace"
version = "0.1.0"
description = ""
authors = ["mlacroix <mlacroix@bol.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.10"
apache-airflow = "2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
